requirements.txt
customtkinter>=5.2.0
tkinterdnd2>=0.3.0
Pillow>=10.0.0
pyinstaller>=6.0.0